requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
web3==6.11.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
cryptography==41.0.7
flasgger==0.9.5
Werkzeug==2.3.7
